Fix S3168: Rule should not be too strict on first parameter name
