+++ sonaranalyzer-dotnet/src/SonarAnalyzer.CSharp/Rules/MethodsShouldUseBaseTypes.cs
@@ -61,7 +61,7 @@ namespace SonarAnalyzer.Rules.CSharp
                 methodSymbol.Parameters.Length == 0 ||
                 methodSymbol.IsOverride ||
                 methodSymbol.GetInterfaceMember() != null ||
-                IsEventHandlerImplementation(methodSymbol))
+                methodSymbol.IsProbablyEventHandler())
             {
                 return Enumerable.Empty<Diagnostic>();
             }
@@ -234,17 +234,6 @@ namespace SonarAnalyzer.Rules.CSharp
             return overridenSymbol != null
                 ? overridenSymbol.ContainingType
                 : accessedMember.ContainingType;
-        }
-
-        private bool IsEventHandlerImplementation(IMethodSymbol methodSymbol)
-        {
-            // Cannot find a way to precisely tell whether this method is a event handler implementation
-            // without doing too much browsing in all trees.
-            // Let's go for a simple heuristic.
-            return methodSymbol.ReturnsVoid &&
-                methodSymbol.Parameters.Length == 2 &&
-                methodSymbol.Parameters[0].Name == "sender" &&
-                methodSymbol.Parameters[0].Type.Is(KnownType.System_Object);
         }
 
         private class ParameterData
