+++ sonaranalyzer-dotnet/src/SonarAnalyzer.Common/Helpers/KnownMethods.cs
@@ -245,6 +245,16 @@ namespace SonarAnalyzer.Helpers
                 methodSymbol.ContainingType.ConstructedFrom.Is(KnownType.System_Collections_Generic_List_T);
         }
 
+        public static bool IsProbablyEventHandler(this IMethodSymbol methodSymbol)
+        {
+            return methodSymbol.ReturnsVoid &&
+                methodSymbol.Parameters.Length == 2 &&
+                methodSymbol.Parameters[0].Type.Is(KnownType.System_Object) &&
+                // Cannot check that the 2nd parameters is or derives from EventArgs as UWP uses other kind of event args
+                // (e.g. ILeavingBackgroundEventArgs)
+                methodSymbol.Parameters[1].Type.ToString().EndsWith("EventArgs", StringComparison.Ordinal);
+        }
+
         private static bool HasExactlyNParameters(this IMethodSymbol methodSymbol, int parametersCount)
         {
             if (methodSymbol.MethodKind == MethodKind.Ordinary)
